+++ src/Credfeto.DotNet.Repo.Formatter/Setup/ApplicationSetup.cs
@@ -1,3 +1,4 @@
+using Credfeto.ChangeLog;
 using Credfeto.DotNet.Repo.Tools.Build;
 using Credfeto.DotNet.Repo.Tools.CleanUp;
 using Microsoft.Extensions.DependencyInjection;
@@ -8,6 +9,6 @@ public static class ApplicationSetup
 {
     public static IServiceCollection AddServices(this IServiceCollection services)
     {
-        return services.AddBuild().AddCleanUp();
+        return services.AddChangeLog().AddBuild().AddCleanUp();
     }
 }